Add files via upload
This android based application is an emergency SOS app which can be used by old age people ,girls and anybody. It triggers an SOS when a button is pressed. When an SOS is triggered message is sent to 5 saved emergency contacts along with current location of the user.As soon as SOS is triggered an audio recording is also started. It uses Google Nearby Places API to search for nearby hospitals, fire stations, and police stations .It also has a fall detect feature ,a loud siren alarm is triggered when either the user or phone falls.